Copied selected regular files to the Sherlock Directory
Introduced Bugs: If the file has already been copied to the Sherlock
directory, an exception is thrown.
If the file is unable to be copied to the Sherlock directory an
exception is thrown. Both need to be handled

Next steps: Directory Processesor must return the extracted files as a
combined collection with regular files to be copied to the Sherlock
directory. Exceptions must be handled